server.py
"""
Degasser/Stripping Tower Design MCP Server.

This server provides packed tower degasser and stripping column design tools including:
- Tier 1: Fast heuristic sizing using Perry's Handbook correlations
- Tier 2: PHREEQC gas-liquid equilibrium modeling
- Tier 3: WaterTAP economic costing and lifecycle analysis

Applications:
- CO2 stripping (alkalinity removal for RO pretreatment)
- H2S stripping (sulfide removal from groundwater)
- VOC stripping (volatile organic compound removal)

Author: Claude AI
"""

import logging
import os
from pathlib import Path
from mcp.server.fastmcp import FastMCP

# Configure logging with debug support
logging.basicConfig(
    level=logging.DEBUG,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    handlers=[
        logging.FileHandler('debug.log'),
        logging.StreamHandler()
    ]
)
logger = logging.getLogger("degasser-design-mcp")

# Initialize the MCP server
mcp = FastMCP("degasser-design-calculator")

# Import core tools
# Phase 1: Tier 1 Heuristic Sizing (COMPLETE)
from tools.heuristic_sizing import heuristic_sizing, list_available_packings

# Phase 2: Tier 2 PHREEQC Gas-Liquid Equilibrium (TO BE IMPLEMENTED)
# from tools.phreeqc_gas_equilibrium import simulate_gas_stripping

# Phase 3: Tier 3 WaterTAP Economic Costing (TO BE IMPLEMENTED)
# from tools.watertap_costing import cost_degasser_system

# Phase 4: Report Generation & Batch Optimization (TO BE IMPLEMENTED)
# from tools.report_generator import generate_degasser_report
# from tools.batch_optimization import batch_optimize_degasser

# Register tools
# Tool 1: Fast Perry's-based heuristic sizing
mcp.tool()(heuristic_sizing)

# Tool 2: List available packing catalog
mcp.tool()(list_available_packings)

# Tool 3: PHREEQC gas-liquid equilibrium (Phase 2)
# mcp.tool()(simulate_gas_stripping)

# Tool 4: WaterTAP economic analysis (Phase 3)
# mcp.tool()(cost_degasser_system)

# Tool 5: Professional HTML reports (Phase 4)
# mcp.tool()(generate_degasser_report)

# Tool 6: Parameter sweeps & optimization (Phase 4)
# mcp.tool()(batch_optimize_degasser)

# Check for required dependencies
try:
    from phreeqpython import PhreeqPython
    PHREEQPYTHON_AVAILABLE = True
    logger.info("PhreeqPython is available")
except ImportError:
    PHREEQPYTHON_AVAILABLE = False
    logger.warning("PhreeqPython not available - PHREEQC tools will not work")

# Check for database files
DB_DIR = Path(__file__).parent / "databases"
DATABASES = {
    "voc_properties": DB_DIR / "voc_properties.json",
    "packing_catalog": DB_DIR / "pack.json",
    "water_properties": DB_DIR / "ppow.json",
    "henrys_law": DB_DIR / "henrys_law.db",
    "phreeqc_voc_phases": DB_DIR / "voc_phases.dat"
}

if __name__ == "__main__":
    logger.info("Starting Degasser Design MCP server...")
    logger.info(f"PhreeqPython available: {PHREEQPYTHON_AVAILABLE}")

    # Check database files
    logger.info("\n=== CHECKING DATABASE FILES ===")
    for db_name, db_path in DATABASES.items():
        if db_path.exists():
            size_mb = db_path.stat().st_size / (1024 * 1024)
            logger.info(f"  ✓ {db_name}: {db_path.name} ({size_mb:.2f} MB)")
        else:
            logger.warning(f"  ✗ {db_name}: {db_path.name} NOT FOUND")

    # Log server status
    logger.info("\n=== DEGASSER DESIGN MCP SERVER ===")
    logger.info("\n📋 THREE-TIER ARCHITECTURE:")
    logger.info("  Tier 1: Fast Heuristic Sizing (<1 sec)")
    logger.info("    - Perry's Handbook correlations")
    logger.info("    - Eckert flooding, HTU/NTU methods")
    logger.info("  Tier 2: PHREEQC Gas-Liquid Equilibrium (10-30 sec)")
    logger.info("    - GAS_PHASE blocks for CO2, H2S, VOC")
    logger.info("    - Multi-stage tower simulation")
    logger.info("  Tier 3: WaterTAP Economic Costing (5-10 sec)")
    logger.info("    - CAPEX, OPEX, LCOW calculations")
    logger.info("    - EPA-WBS correlations")

    logger.info("\n🎯 APPLICATIONS:")
    logger.info("  1. CO2 Stripping (Alkalinity Removal)")
    logger.info("     - RO pretreatment, boiler feedwater")
    logger.info("     - Air/water: 20:1-50:1, pH: 4.5-5.5")
    logger.info("  2. H2S Stripping (Sulfide Removal)")
    logger.info("     - Groundwater, industrial wastewater")
    logger.info("     - Air/water: 30:1-100:1, pH: 4.0-5.0")
    logger.info("  3. VOC Stripping (Volatile Organic Removal)")
    logger.info("     - Contaminated groundwater remediation")
    logger.info("     - Henry's law governed mass transfer")

    logger.info("\n📊 DATA SOURCES:")
    logger.info("  • Perry's Chemical Engineers' Handbook (semantic search)")
    logger.info("  • henrys-law.org SQLite database (2 MB, 4632 compounds)")
    logger.info("  • VOC properties from Air-stripping-column repo")
    logger.info("  • Packing catalog with Eckert correlation data")

    logger.info("\n✅ IMPLEMENTATION STATUS: PHASE 1 COMPLETE")
    logger.info("  ✅ Phase 0: Data Acquisition")
    logger.info("     - Downloaded databases from GitHub")
    logger.info("     - Generated unified VOC properties")
    logger.info("     - Created PHREEQC phases definitions")
    logger.info("  ✅ Phase 1: Tier 1 Heuristic Sizing (COMPLETE)")
    logger.info("     - Perry's Eckert GPDC flooding correlation")
    logger.info("     - HTU/NTU method with Eq 14-158")
    logger.info("     - 9 packings in catalog with actual properties")
    logger.info("     - MCP tools: heuristic_sizing, list_available_packings")
    logger.info("  ⏳ Phase 2: Tier 2 PHREEQC Simulation (PENDING)")
    logger.info("  ⏳ Phase 3: Tier 3 WaterTAP Costing (PENDING)")
    logger.info("  ⏳ Phase 4: Reports & Optimization (PENDING)")

    logger.info("\n=== SERVER READY FOR DEVELOPMENT ===")

    # Start the server
    mcp.run()
